+++ apps/docs/src/guide/transactions/snippets/assemble-tx/assemble-tx-1.ts
@@ -1,30 +1,33 @@
 import { bn, Provider, ScriptTransactionRequest, Wallet } from 'fuels';
 import { TestAssetId } from 'fuels/test-utils';
 
-import { LOCAL_NETWORK_URL, WALLET_PVT_KEY } from '../../../env';
+import { LOCAL_NETWORK_URL, WALLET_PVT_KEY } from '../../../../env';
 
 const provider = new Provider(LOCAL_NETWORK_URL);
 const account = Wallet.fromPrivateKey(WALLET_PVT_KEY, provider);
 
 const baseAssetId = await provider.getBaseAssetId();
 
-const signatureCallback = async (request: ScriptTransactionRequest) =>
-  Promise.resolve(request);
-
-// #region estimate-and-fund-1
+// #region assemble-tx-1
 const request = new ScriptTransactionRequest();
 
-await request.estimateAndFund(account, {
-  signatureCallback,
-  quantities: [
+const { assembledRequest } = await provider.assembleTx({
+  request,
+  feePayerAccount: account,
+  accountCoinQuantities: [
     {
       amount: bn(100),
       assetId: baseAssetId,
+      account,
+      changeOutputAccount: account,
     },
     {
       amount: bn(200),
       assetId: TestAssetId.A.value,
+      account,
+      changeOutputAccount: account,
     },
   ],
 });
-// #endregion estimate-and-fund-1
+// #endregion assemble-tx-1
+console.log(assembledRequest, 'assembledRequest');
