+++ apps/docs/src/guide/transactions/snippets/assemble-tx/change-output-collision.ts
@@ -5,7 +5,7 @@ import {
   LOCAL_NETWORK_URL,
   WALLET_PVT_KEY,
   WALLET_PVT_KEY_2,
-} from '../../../env';
+} from '../../../../env';
 
 const provider = new Provider(LOCAL_NETWORK_URL);
 const baseAssetId = await provider.getBaseAssetId();
